requirements.txt
google-adk
google-genai
python-dotenv==1.0.1
pydantic==2.10.6
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
